refactor: initialize app
