--- packages/plugin-core/src/generator/templates/app/react/runApp.tsx.ejs
@@ -1,27 +1,0 @@
-import { createBaseApp, emitLifeCycles, staticConfig, createHistory } from './common';
-<% if(isMiniapp){ %>
-  import miniappRenderer from 'renderer-miniapp';
-<% } else { %>
-  import reactRenderer from 'renderer-react';
-<% } %>
-
-<% if (globalStyle) {%>
-  import '../<%= globalStyle %>'
-<% } %>
-
-export function runApp(appConfig) {
-  <% if(isMiniapp){ %>
-    miniappRenderer({
-      appConfig,
-      createBaseApp,
-      staticConfig,
-      createHistory
-    });
-  <% } else { %>
-    reactRenderer({
-      appConfig,
-      emitLifeCycles,
-      createBaseApp,
-    });
-  <% } %>
-};
