+++ packages/renderer-miniapp/src/miniappRenderer.ts
@@ -1,9 +1,9 @@
 import { isMiniApp, isWeChatMiniProgram, isByteDanceMicroApp } from 'universal-env';
 
-function renderer({ appConfig, createBaseApp, createHistory, staticConfig }) {
+function miniappRenderer({ appConfig, createBaseApp, createHistory, staticConfig }) {
   const env = { isMiniApp, isWeChatMiniProgram, isByteDanceMicroApp };
   createBaseApp(appConfig, {}, { env });
   createHistory({ routes: staticConfig.routes });
 };
 
-export default renderer;
+export default miniappRenderer;
