+++ packages/renderer-rax/README.md
@@ -1,11 +1,1 @@
 # `renderer-rax`
-
-> TODO: description
-
-## Usage
-
-```
-const rendererRax = require('renderer-rax');
-
-// TODO: DEMONSTRATE API
-```
